requirements.txt.txt
pandas
streamlit
textblob
langchain
openai
faiss-cpu
tiktoken
python-dotenv
